requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
geopy>=2.3.0